+++ src/components/AssignmentsView.vue
@@ -124,11 +124,15 @@ export default {
             .then(body => {
                 const assignments = body
                 //Filter assignments that are not due, only get ones that are unsubmitted
-                .filter(assignment => assignment.due_at && assignment.submission.workflow_state === "unsubmitted")
+                //.filter(assignment => assignment.due_at && assignment.submission.workflow_state === "unsubmitted")
                 //Pre-compute assignment information
                 .map(assignment => {
                     if(!(assignment.id in this.checkedAssignments))
                         this.$set(this.checkedAssignments, assignment.id, false)
+                    if(assignment.submission.workflow_state === "graded" || assignment.submission.workflow_state === "submitted") {
+                        console.debug('assignment', assignment.id, 'is submitted')
+                        this.$set(this.checkedAssignments, assignment.id, true)
+                    }
                     const date = new Date(assignment.due_at)
                     const delta = date.valueOf() - NOW;