+++ src/views/Home.vue
@@ -41,9 +41,9 @@
         <Clock />
         <div class="box">
           <div class="buttons is-centered">
-            <b-button @click="openScheduler" type="is-primary" icon-left="calendar">Manage Schedule</b-button>
-            <b-button @click="openCourseManager" type="is-dark" icon-left="school">Manage Courses</b-button>
-            <b-button @click="openSync" type="is-info" icon-left="export">Export / Import</b-button>
+            <b-button @click="openScheduler" type="is-primary" icon-left="calendar">Schedule</b-button>
+            <b-button @click="openCourseManager" type="is-dark" icon-left="school">Courses</b-button>
+            <b-button @click="openSync" type="is-info" icon-left="sync">Sync</b-button>
             <!--<b-button type="is-info" icon-left="sync" disabled>Sync</b-button>-->
           </div>
         </div>
@@ -159,6 +159,7 @@ export default {
           },
           events: {
               submit: (schedule) => {
+                  console.info('Recieved new schedule', schedule)
                   this.schedule = schedule
                   window.localStorage.canvas_schedule = JSON.stringify(schedule)
               }